setting an icon works, full functionality acheived
