+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
             </intent-filter>
         </activity>
 
-        <activity android:name=".main.timer.CircuitCreate" />
+        <activity android:name=".main.circuit.CircuitCreate" />
 
         <activity android:name=".main.MainActivity" />
     </application>
