+++ app/src/main/res/drawable/ic_boxer.xml
@@ -1,6 +1,6 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="70dp"
-    android:height="70dp"
+    android:width="50dp"
+    android:height="50dp"
     android:viewportWidth="512"
     android:viewportHeight="512">
     <path
