+++ app/src/main/res/drawable/ic_dumbbell.xml
@@ -1,6 +1,6 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="70dp"
-    android:height="70dp"
+    android:width="50dp"
+    android:height="50dp"
     android:viewportWidth="511.984"
     android:viewportHeight="511.984">
   <path
