+++ app/src/main/res/drawable/ic_gym_4.xml
@@ -1,6 +1,6 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="70dp"
-    android:height="70dp"
+    android:width="50dp"
+    android:height="50dp"
     android:viewportWidth="512.009"
     android:viewportHeight="512.009">
   <path
