+++ app/src/main/res/drawable/ic_gymnast.xml
@@ -1,12 +1,12 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="70dp"
-    android:height="70dp"
+    android:width="50dp"
+    android:height="50dp"
     android:viewportWidth="512"
     android:viewportHeight="512">
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M170.667,213.333c-70.583,0 -128,57.417 -128,128s57.417,128 128,128s128,-57.417 128,-128S241.25,213.333 170.667,213.333zM170.667,448C111.854,448 64,400.146 64,341.333c0,-58.813 47.854,-106.667 106.667,-106.667c4.156,0 8.103,0.762 12.137,1.227c-7.923,19.924 -12.137,40.934 -12.137,62.773c0,50.333 22.577,97.896 60.708,130.172C214.117,440.852 193.24,448 170.667,448zM247.736,414.685C212.79,386.438 192,343.833 192,298.667c0,-20.302 4.19,-39.766 11.999,-58.138c6.835,2.27 13.32,5.148 19.466,8.686c-6.533,15.605 -10.132,32.456 -10.132,49.452c0,38.697 17.647,74.961 47.432,99.181C256.961,403.892 252.638,409.535 247.736,414.685zM234.667,298.667c0,-12.745 2.216,-25.044 6.52,-36.741c21.995,19.555 36.147,47.734 36.147,79.408c0,13.034 -2.682,25.378 -6.982,36.934C247.819,358.142 234.667,329.263 234.667,298.667z"/>
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M405.333,140.565V128c0,-19.826 -13.654,-36.382 -32,-41.151V10.667C373.333,4.771 368.563,0 362.667,0H320c-5.896,0 -10.667,4.771 -10.667,10.667v76.182c-18.346,4.77 -32,21.326 -32,41.151v12.694c-15.504,6.283 -30.016,14.841 -43.128,25.409c-1.922,-17.759 -14.639,-32.194 -31.539,-36.587V10.667C202.667,4.771 197.896,0 192,0h-42.667c-5.896,0 -10.667,4.771 -10.667,10.667v118.849c-18.346,4.77 -32,21.326 -32,41.151v12.565C44.203,208.613 0,269.867 0,341.333C0,435.438 76.563,512 170.667,512c47.262,0 90.079,-19.329 121.016,-50.477c16.064,4.896 32.659,7.81 49.651,7.81c94.104,0 170.667,-76.563 170.667,-170.667C512,227.201 467.797,165.947 405.333,140.565zM330.667,21.333H352v64h-21.333V21.333zM298.667,128c0,-11.76 9.563,-21.333 21.333,-21.333h42.667C374.438,106.667 384,116.24 384,128v5.617c-13.665,-3.538 -27.911,-5.617 -42.667,-5.617c-14.522,0 -28.813,2.085 -42.667,5.681V128zM160,21.333h21.333V128H160V21.333zM128,170.667c0,-11.76 9.563,-21.333 21.333,-21.333H192c11.771,0 21.333,9.573 21.333,21.333v5.617c-13.665,-3.538 -27.911,-5.617 -42.667,-5.617c-14.755,0 -29.001,2.079 -42.667,5.617V170.667zM170.667,490.667c-82.333,0 -149.333,-66.99 -149.333,-149.333S88.333,192 170.667,192S320,258.99 320,341.333S253,490.667 170.667,490.667zM280.88,211.249C298.612,198.979 319.513,192 341.333,192C400.146,192 448,239.854 448,298.667s-47.854,106.667 -106.667,106.667c-4.12,0 -8.228,-0.254 -12.301,-0.721c7.858,-19.592 12.301,-40.911 12.301,-63.279C341.333,289.233 317.801,242.579 280.88,211.249zM341.333,448c-11.686,0 -23.094,-1.823 -34.307,-4.47c4.539,-6.042 8.762,-12.323 12.48,-18.945c7.195,1.24 14.48,2.082 21.827,2.082c70.583,0 128,-57.417 128,-128s-57.417,-128 -128,-128c-28.595,0 -55.922,9.927 -78.31,27.365c-6.543,-4.232 -13.411,-7.975 -20.526,-11.296c27.141,-23.928 62.379,-37.402 98.836,-37.402c82.333,0 149.333,66.99 149.333,149.333S423.667,448 341.333,448z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M170.667,213.333c-70.583,0 -128,57.417 -128,128s57.417,128 128,128s128,-57.417 128,-128S241.25,213.333 170.667,213.333zM170.667,448C111.854,448 64,400.146 64,341.333c0,-58.813 47.854,-106.667 106.667,-106.667c4.156,0 8.103,0.762 12.137,1.227c-7.923,19.924 -12.137,40.934 -12.137,62.773c0,50.333 22.577,97.896 60.708,130.172C214.117,440.852 193.24,448 170.667,448zM247.736,414.685C212.79,386.438 192,343.833 192,298.667c0,-20.302 4.19,-39.766 11.999,-58.138c6.835,2.27 13.32,5.148 19.466,8.686c-6.533,15.605 -10.132,32.456 -10.132,49.452c0,38.697 17.647,74.961 47.432,99.181C256.961,403.892 252.638,409.535 247.736,414.685zM234.667,298.667c0,-12.745 2.216,-25.044 6.52,-36.741c21.995,19.555 36.147,47.734 36.147,79.408c0,13.034 -2.682,25.378 -6.982,36.934C247.819,358.142 234.667,329.263 234.667,298.667z" />
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M405.333,140.565V128c0,-19.826 -13.654,-36.382 -32,-41.151V10.667C373.333,4.771 368.563,0 362.667,0H320c-5.896,0 -10.667,4.771 -10.667,10.667v76.182c-18.346,4.77 -32,21.326 -32,41.151v12.694c-15.504,6.283 -30.016,14.841 -43.128,25.409c-1.922,-17.759 -14.639,-32.194 -31.539,-36.587V10.667C202.667,4.771 197.896,0 192,0h-42.667c-5.896,0 -10.667,4.771 -10.667,10.667v118.849c-18.346,4.77 -32,21.326 -32,41.151v12.565C44.203,208.613 0,269.867 0,341.333C0,435.438 76.563,512 170.667,512c47.262,0 90.079,-19.329 121.016,-50.477c16.064,4.896 32.659,7.81 49.651,7.81c94.104,0 170.667,-76.563 170.667,-170.667C512,227.201 467.797,165.947 405.333,140.565zM330.667,21.333H352v64h-21.333V21.333zM298.667,128c0,-11.76 9.563,-21.333 21.333,-21.333h42.667C374.438,106.667 384,116.24 384,128v5.617c-13.665,-3.538 -27.911,-5.617 -42.667,-5.617c-14.522,0 -28.813,2.085 -42.667,5.681V128zM160,21.333h21.333V128H160V21.333zM128,170.667c0,-11.76 9.563,-21.333 21.333,-21.333H192c11.771,0 21.333,9.573 21.333,21.333v5.617c-13.665,-3.538 -27.911,-5.617 -42.667,-5.617c-14.755,0 -29.001,2.079 -42.667,5.617V170.667zM170.667,490.667c-82.333,0 -149.333,-66.99 -149.333,-149.333S88.333,192 170.667,192S320,258.99 320,341.333S253,490.667 170.667,490.667zM280.88,211.249C298.612,198.979 319.513,192 341.333,192C400.146,192 448,239.854 448,298.667s-47.854,106.667 -106.667,106.667c-4.12,0 -8.228,-0.254 -12.301,-0.721c7.858,-19.592 12.301,-40.911 12.301,-63.279C341.333,289.233 317.801,242.579 280.88,211.249zM341.333,448c-11.686,0 -23.094,-1.823 -34.307,-4.47c4.539,-6.042 8.762,-12.323 12.48,-18.945c7.195,1.24 14.48,2.082 21.827,2.082c70.583,0 128,-57.417 128,-128s-57.417,-128 -128,-128c-28.595,0 -55.922,9.927 -78.31,27.365c-6.543,-4.232 -13.411,-7.975 -20.526,-11.296c27.141,-23.928 62.379,-37.402 98.836,-37.402c82.333,0 149.333,66.99 149.333,149.333S423.667,448 341.333,448z" />
 </vector>
