+++ app/src/main/res/drawable/ic_jump_rope.xml
@@ -1,9 +1,9 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="70dp"
-    android:height="70dp"
+    android:width="50dp"
+    android:height="50dp"
     android:viewportWidth="512"
     android:viewportHeight="512">
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M437.333,306.99V288c0,-17.646 -14.354,-32 -32,-32V87.854C405.333,39.406 365.927,0 317.479,0h-3.115c-28.979,0 -55.427,13.938 -71.729,37.104c-14.125,-10.135 -31.323,-15.771 -48.802,-15.771h-3.667c-46.042,0 -83.5,37.458 -83.5,83.5V256c-17.646,0 -32,14.354 -32,32v18.99C68.168,312.852 64,321.246 64,330.667v128C64,488.073 87.927,512 117.333,512s53.333,-23.927 53.333,-53.333v-128c0,-9.421 -4.168,-17.815 -10.667,-23.677V288c0,-17.646 -14.354,-32 -32,-32V104.833c0,-34.281 27.885,-62.167 62.167,-62.167h3.667c12.115,0 24.018,3.673 34.091,10.258C203.21,71.93 149.333,118.143 149.333,160c0,47.448 34.99,74.667 96,74.667c61.01,0 96,-27.219 96,-74.667c0,-43.901 -59.194,-92.53 -81.917,-109.594c12.284,-18.112 32.609,-29.073 54.948,-29.073h3.115c36.677,0 66.521,29.844 66.521,66.521V256c-17.646,0 -32,14.354 -32,32v18.99c-6.499,5.862 -10.667,14.257 -10.667,23.677v128c0,29.406 23.927,53.333 53.333,53.333S448,488.073 448,458.667v-128C448,321.246 443.832,312.852 437.333,306.99zM149.333,330.667v128c0,17.646 -14.354,32 -32,32c-17.646,0 -32,-14.354 -32,-32v-128C85.333,324.781 90.115,320 96,320h42.667C144.552,320 149.333,324.781 149.333,330.667zM138.667,288v10.667H96V288c0,-5.885 4.781,-10.667 10.667,-10.667H128C133.885,277.333 138.667,282.115 138.667,288zM320,160c0,35.385 -25.125,53.333 -74.667,53.333c-49.542,0 -74.667,-17.948 -74.667,-53.333c0,-31.281 49.26,-74.521 74.667,-93.5C270.75,85.458 320,128.667 320,160zM373.333,288c0,-5.885 4.781,-10.667 10.667,-10.667h21.333c5.885,0 10.667,4.781 10.667,10.667v10.667h-42.667V288zM426.667,458.667c0,17.646 -14.354,32 -32,32s-32,-14.354 -32,-32v-128c0,-5.885 4.781,-10.667 10.667,-10.667H416c5.885,0 10.667,4.781 10.667,10.667V458.667z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M437.333,306.99V288c0,-17.646 -14.354,-32 -32,-32V87.854C405.333,39.406 365.927,0 317.479,0h-3.115c-28.979,0 -55.427,13.938 -71.729,37.104c-14.125,-10.135 -31.323,-15.771 -48.802,-15.771h-3.667c-46.042,0 -83.5,37.458 -83.5,83.5V256c-17.646,0 -32,14.354 -32,32v18.99C68.168,312.852 64,321.246 64,330.667v128C64,488.073 87.927,512 117.333,512s53.333,-23.927 53.333,-53.333v-128c0,-9.421 -4.168,-17.815 -10.667,-23.677V288c0,-17.646 -14.354,-32 -32,-32V104.833c0,-34.281 27.885,-62.167 62.167,-62.167h3.667c12.115,0 24.018,3.673 34.091,10.258C203.21,71.93 149.333,118.143 149.333,160c0,47.448 34.99,74.667 96,74.667c61.01,0 96,-27.219 96,-74.667c0,-43.901 -59.194,-92.53 -81.917,-109.594c12.284,-18.112 32.609,-29.073 54.948,-29.073h3.115c36.677,0 66.521,29.844 66.521,66.521V256c-17.646,0 -32,14.354 -32,32v18.99c-6.499,5.862 -10.667,14.257 -10.667,23.677v128c0,29.406 23.927,53.333 53.333,53.333S448,488.073 448,458.667v-128C448,321.246 443.832,312.852 437.333,306.99zM149.333,330.667v128c0,17.646 -14.354,32 -32,32c-17.646,0 -32,-14.354 -32,-32v-128C85.333,324.781 90.115,320 96,320h42.667C144.552,320 149.333,324.781 149.333,330.667zM138.667,288v10.667H96V288c0,-5.885 4.781,-10.667 10.667,-10.667H128C133.885,277.333 138.667,282.115 138.667,288zM320,160c0,35.385 -25.125,53.333 -74.667,53.333c-49.542,0 -74.667,-17.948 -74.667,-53.333c0,-31.281 49.26,-74.521 74.667,-93.5C270.75,85.458 320,128.667 320,160zM373.333,288c0,-5.885 4.781,-10.667 10.667,-10.667h21.333c5.885,0 10.667,4.781 10.667,10.667v10.667h-42.667V288zM426.667,458.667c0,17.646 -14.354,32 -32,32s-32,-14.354 -32,-32v-128c0,-5.885 4.781,-10.667 10.667,-10.667H416c5.885,0 10.667,4.781 10.667,10.667V458.667z" />
 </vector>
