+++ app/src/main/res/drawable/ic_mat.xml
@@ -1,15 +1,15 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="70dp"
-    android:height="70dp"
+    android:width="50dp"
+    android:height="50dp"
     android:viewportWidth="512.034"
     android:viewportHeight="512.034">
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M469.35,85.367H191.423c-2.594,-23.052 -13.51,-44.615 -30.833,-60.375c-20.26,-18.427 -46.635,-27.219 -73.969,-24.51C38.058,5.117 0.017,48.7 0.017,99.7v316.333h0.021c0.031,21.698 7.229,42.688 20.885,59.813c18.333,23 45.708,36.188 75.094,36.188H469.35c23.531,0 42.667,-19.135 42.667,-42.667V128.034C512.017,104.502 492.881,85.367 469.35,85.367zM21.35,99.7c0,-40.125 29.563,-74.375 67.292,-77.99c21.396,-1.938 41.844,4.74 57.594,19.063c15.531,14.135 24.448,34.271 24.448,55.26l0.365,260.115c-18.542,-23.219 -46.052,-36.094 -75.135,-36.094c-7.396,0 -14.896,0.833 -22.375,2.542c-20.823,4.75 -39.094,16.49 -52.188,32.583V99.7zM490.683,469.367c0,11.76 -9.573,21.333 -21.333,21.333H96.017c-22.854,0 -44.146,-10.26 -58.417,-28.156c-14.271,-17.896 -19.458,-41.313 -14.229,-64.24c6.104,-26.74 28.177,-48.813 54.917,-54.917c5.865,-1.333 11.74,-1.99 17.542,-1.99c22.698,0 44.135,10.01 58.552,28.073c5.573,6.948 14.635,9.5 23.135,6.521c8.677,-3.042 14.5,-11.198 14.5,-20.292v-84.583l195.125,195.125c2.083,2.083 4.813,3.125 7.542,3.125s5.458,-1.042 7.542,-3.125c4.167,-4.167 4.167,-10.917 0,-15.083L192.017,240.95V106.7H469.35c11.76,0 21.333,9.573 21.333,21.333V469.367z"/>
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M423.558,131.159c-4.167,-4.167 -10.917,-4.167 -15.083,0c-4.167,4.167 -4.167,10.917 0,15.083l42.667,42.667c2.083,2.083 4.813,3.125 7.542,3.125s5.458,-1.042 7.542,-3.125c4.167,-4.167 4.167,-10.917 0,-15.083L423.558,131.159z"/>
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M252.892,131.159c-4.167,-4.167 -10.917,-4.167 -15.083,0c-4.167,4.167 -4.167,10.917 0,15.083l213.333,213.333c2.083,2.083 4.813,3.125 7.542,3.125s5.458,-1.042 7.542,-3.125c4.167,-4.167 4.167,-10.917 0,-15.083L252.892,131.159z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M469.35,85.367H191.423c-2.594,-23.052 -13.51,-44.615 -30.833,-60.375c-20.26,-18.427 -46.635,-27.219 -73.969,-24.51C38.058,5.117 0.017,48.7 0.017,99.7v316.333h0.021c0.031,21.698 7.229,42.688 20.885,59.813c18.333,23 45.708,36.188 75.094,36.188H469.35c23.531,0 42.667,-19.135 42.667,-42.667V128.034C512.017,104.502 492.881,85.367 469.35,85.367zM21.35,99.7c0,-40.125 29.563,-74.375 67.292,-77.99c21.396,-1.938 41.844,4.74 57.594,19.063c15.531,14.135 24.448,34.271 24.448,55.26l0.365,260.115c-18.542,-23.219 -46.052,-36.094 -75.135,-36.094c-7.396,0 -14.896,0.833 -22.375,2.542c-20.823,4.75 -39.094,16.49 -52.188,32.583V99.7zM490.683,469.367c0,11.76 -9.573,21.333 -21.333,21.333H96.017c-22.854,0 -44.146,-10.26 -58.417,-28.156c-14.271,-17.896 -19.458,-41.313 -14.229,-64.24c6.104,-26.74 28.177,-48.813 54.917,-54.917c5.865,-1.333 11.74,-1.99 17.542,-1.99c22.698,0 44.135,10.01 58.552,28.073c5.573,6.948 14.635,9.5 23.135,6.521c8.677,-3.042 14.5,-11.198 14.5,-20.292v-84.583l195.125,195.125c2.083,2.083 4.813,3.125 7.542,3.125s5.458,-1.042 7.542,-3.125c4.167,-4.167 4.167,-10.917 0,-15.083L192.017,240.95V106.7H469.35c11.76,0 21.333,9.573 21.333,21.333V469.367z" />
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M423.558,131.159c-4.167,-4.167 -10.917,-4.167 -15.083,0c-4.167,4.167 -4.167,10.917 0,15.083l42.667,42.667c2.083,2.083 4.813,3.125 7.542,3.125s5.458,-1.042 7.542,-3.125c4.167,-4.167 4.167,-10.917 0,-15.083L423.558,131.159z" />
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M252.892,131.159c-4.167,-4.167 -10.917,-4.167 -15.083,0c-4.167,4.167 -4.167,10.917 0,15.083l213.333,213.333c2.083,2.083 4.813,3.125 7.542,3.125s5.458,-1.042 7.542,-3.125c4.167,-4.167 4.167,-10.917 0,-15.083L252.892,131.159z" />
 </vector>
