+++ app/src/main/res/drawable/ic_punching_ball.xml
@@ -1,9 +1,9 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="70dp"
-    android:height="70dp"
+    android:width="50dp"
+    android:height="50dp"
     android:viewportWidth="469.333"
     android:viewportHeight="469.333">
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M466.208,451.125l-56.458,-56.458l6.25,-6.25c8.313,-8.313 8.313,-21.854 0,-30.156l-19.941,-19.98c19.281,-29.927 30.608,-65.444 30.608,-103.613c0,-105.865 -86.125,-192 -192,-192c-38.174,0 -73.693,11.328 -103.621,30.612l-19.962,-19.945c-8.333,-8.313 -21.833,-8.313 -30.167,0l-6.25,6.25L18.208,3.125c-4.167,-4.167 -10.917,-4.167 -15.083,0c-4.167,4.167 -4.167,10.917 0,15.083l56.458,56.458l-6.25,6.25c-8.313,8.313 -8.313,21.854 -0.021,30.156l19.941,20.014c-19.268,29.921 -30.587,65.423 -30.587,103.579c0,105.865 86.125,192 192,192c38.18,0 73.702,-11.331 103.632,-30.618L358.25,416c4.167,4.156 9.625,6.24 15.083,6.24s10.917,-2.083 15.083,-6.24l6.25,-6.25l56.458,56.458c2.083,2.083 4.813,3.125 7.542,3.125c2.729,0 5.458,-1.042 7.542,-3.125C470.375,462.042 470.375,455.292 466.208,451.125zM234.667,64c94.104,0 170.667,76.563 170.667,170.667c0,30.885 -8.372,59.8 -22.78,84.819c-5.77,-41.148 -36.927,-95.609 -87.012,-145.694c-50.095,-50.082 -104.546,-81.247 -145.697,-87.01C174.866,72.372 203.78,64 234.667,64zM68.417,96L96,68.417l17.496,17.482c-10.113,8.253 -19.363,17.508 -27.615,27.622L68.417,96zM234.667,405.333C140.562,405.333 64,328.771 64,234.667c0,-30.885 8.372,-59.799 22.78,-84.819c5.77,41.148 36.927,95.609 87.012,145.694c50.148,50.139 104.629,81.232 145.792,86.952C294.54,396.936 265.594,405.333 234.667,405.333zM188.875,280.458c-70.979,-70.99 -96.333,-144.594 -74.667,-166.25c5,-5.01 12.792,-7.5 22.688,-7.5c32.875,0 88.979,27.594 143.562,82.167c70.979,70.99 96.333,144.594 74.667,166.25C333.479,376.771 259.896,351.448 188.875,280.458zM373.333,400.917l-17.488,-17.488c10.107,-8.247 19.352,-17.496 27.599,-27.603l17.473,17.508L373.333,400.917z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M466.208,451.125l-56.458,-56.458l6.25,-6.25c8.313,-8.313 8.313,-21.854 0,-30.156l-19.941,-19.98c19.281,-29.927 30.608,-65.444 30.608,-103.613c0,-105.865 -86.125,-192 -192,-192c-38.174,0 -73.693,11.328 -103.621,30.612l-19.962,-19.945c-8.333,-8.313 -21.833,-8.313 -30.167,0l-6.25,6.25L18.208,3.125c-4.167,-4.167 -10.917,-4.167 -15.083,0c-4.167,4.167 -4.167,10.917 0,15.083l56.458,56.458l-6.25,6.25c-8.313,8.313 -8.313,21.854 -0.021,30.156l19.941,20.014c-19.268,29.921 -30.587,65.423 -30.587,103.579c0,105.865 86.125,192 192,192c38.18,0 73.702,-11.331 103.632,-30.618L358.25,416c4.167,4.156 9.625,6.24 15.083,6.24s10.917,-2.083 15.083,-6.24l6.25,-6.25l56.458,56.458c2.083,2.083 4.813,3.125 7.542,3.125c2.729,0 5.458,-1.042 7.542,-3.125C470.375,462.042 470.375,455.292 466.208,451.125zM234.667,64c94.104,0 170.667,76.563 170.667,170.667c0,30.885 -8.372,59.8 -22.78,84.819c-5.77,-41.148 -36.927,-95.609 -87.012,-145.694c-50.095,-50.082 -104.546,-81.247 -145.697,-87.01C174.866,72.372 203.78,64 234.667,64zM68.417,96L96,68.417l17.496,17.482c-10.113,8.253 -19.363,17.508 -27.615,27.622L68.417,96zM234.667,405.333C140.562,405.333 64,328.771 64,234.667c0,-30.885 8.372,-59.799 22.78,-84.819c5.77,41.148 36.927,95.609 87.012,145.694c50.148,50.139 104.629,81.232 145.792,86.952C294.54,396.936 265.594,405.333 234.667,405.333zM188.875,280.458c-70.979,-70.99 -96.333,-144.594 -74.667,-166.25c5,-5.01 12.792,-7.5 22.688,-7.5c32.875,0 88.979,27.594 143.562,82.167c70.979,70.99 96.333,144.594 74.667,166.25C333.479,376.771 259.896,351.448 188.875,280.458zM373.333,400.917l-17.488,-17.488c10.107,-8.247 19.352,-17.496 27.599,-27.603l17.473,17.508L373.333,400.917z" />
 </vector>
