+++ app/src/main/res/drawable/ic_resistance_1.xml
@@ -1,9 +1,9 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="70dp"
-    android:height="70dp"
+    android:width="50dp"
+    android:height="50dp"
     android:viewportWidth="512"
     android:viewportHeight="512">
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M500.042,420.958l-67.125,-67.125c-7.938,-7.948 -18.938,-12.5 -30.167,-12.5h-29.417v-41.146c0,-30.156 -17.313,-56.052 -45.167,-67.594c-27.813,-11.531 -58.375,-5.479 -79.708,15.865c-18.021,18.031 -41.021,17.656 -56.458,11.24c-15.479,-6.406 -32,-22.354 -32,-47.885v-41.146h29.417c11.229,0 22.229,-4.552 30.167,-12.5l67.125,-67.135c7.604,-7.604 11.958,-18.125 11.958,-28.865c0,-22.521 -18.313,-40.833 -40.833,-40.833h-3.798C249.617,8.944 237.888,0 224,0H74.667C60.779,0 49.049,8.944 44.632,21.333h-3.798C18.313,21.333 0,39.646 0,62.167c0,10.74 4.354,21.26 11.958,28.875l67.125,67.125c7.938,7.948 18.938,12.5 30.167,12.5h29.417v41.146c0,30.156 17.313,56.052 45.167,67.594c27.833,11.51 58.396,5.479 79.708,-15.865c18.021,-18.031 41.021,-17.646 56.458,-11.24c15.479,6.406 32,22.354 32,47.885v41.146h-29.417c-11.229,0 -22.229,4.552 -30.167,12.5l-67.125,67.135c-7.604,7.604 -11.958,18.125 -11.958,28.865c0,22.521 18.313,40.833 40.833,40.833h3.798C262.383,503.056 274.112,512 288,512h149.333c13.888,0 25.617,-8.944 30.035,-21.333h3.798c22.521,0 40.833,-18.313 40.833,-40.833C512,439.094 507.646,428.573 500.042,420.958zM74.667,21.333H224c5.875,0 10.667,4.781 10.667,10.667S229.875,42.667 224,42.667H74.667C68.792,42.667 64,37.885 64,32S68.792,21.333 74.667,21.333zM109.25,149.333c-5.604,0 -11.104,-2.281 -15.083,-6.25L27.042,75.969c-3.625,-3.646 -5.708,-8.677 -5.708,-13.802c0,-10.75 8.75,-19.5 19.5,-19.5h3.798C49.049,55.056 60.779,64 74.667,64H224c13.888,0 25.617,-8.944 30.035,-21.333h3.798c10.75,0 19.5,8.75 19.5,19.5c0,5.125 -2.083,10.156 -5.708,13.792L204.5,143.083c-3.979,3.969 -9.479,6.25 -15.083,6.25H109.25zM437.333,490.667H288c-5.875,0 -10.667,-4.781 -10.667,-10.667c0,-5.885 4.792,-10.667 10.667,-10.667h149.333c5.875,0 10.667,4.781 10.667,10.667C448,485.885 443.208,490.667 437.333,490.667zM471.167,469.333h-3.798C462.951,456.944 451.221,448 437.333,448H288c-13.888,0 -25.617,8.944 -30.035,21.333h-3.798c-10.75,0 -19.5,-8.75 -19.5,-19.5c0,-5.125 2.083,-10.156 5.708,-13.792l67.125,-67.125c3.979,-3.969 9.479,-6.25 15.083,-6.25h80.167c5.604,0 11.104,2.281 15.083,6.25l67.125,67.115c3.625,3.646 5.708,8.677 5.708,13.802C490.667,460.583 481.917,469.333 471.167,469.333z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M500.042,420.958l-67.125,-67.125c-7.938,-7.948 -18.938,-12.5 -30.167,-12.5h-29.417v-41.146c0,-30.156 -17.313,-56.052 -45.167,-67.594c-27.813,-11.531 -58.375,-5.479 -79.708,15.865c-18.021,18.031 -41.021,17.656 -56.458,11.24c-15.479,-6.406 -32,-22.354 -32,-47.885v-41.146h29.417c11.229,0 22.229,-4.552 30.167,-12.5l67.125,-67.135c7.604,-7.604 11.958,-18.125 11.958,-28.865c0,-22.521 -18.313,-40.833 -40.833,-40.833h-3.798C249.617,8.944 237.888,0 224,0H74.667C60.779,0 49.049,8.944 44.632,21.333h-3.798C18.313,21.333 0,39.646 0,62.167c0,10.74 4.354,21.26 11.958,28.875l67.125,67.125c7.938,7.948 18.938,12.5 30.167,12.5h29.417v41.146c0,30.156 17.313,56.052 45.167,67.594c27.833,11.51 58.396,5.479 79.708,-15.865c18.021,-18.031 41.021,-17.646 56.458,-11.24c15.479,6.406 32,22.354 32,47.885v41.146h-29.417c-11.229,0 -22.229,4.552 -30.167,12.5l-67.125,67.135c-7.604,7.604 -11.958,18.125 -11.958,28.865c0,22.521 18.313,40.833 40.833,40.833h3.798C262.383,503.056 274.112,512 288,512h149.333c13.888,0 25.617,-8.944 30.035,-21.333h3.798c22.521,0 40.833,-18.313 40.833,-40.833C512,439.094 507.646,428.573 500.042,420.958zM74.667,21.333H224c5.875,0 10.667,4.781 10.667,10.667S229.875,42.667 224,42.667H74.667C68.792,42.667 64,37.885 64,32S68.792,21.333 74.667,21.333zM109.25,149.333c-5.604,0 -11.104,-2.281 -15.083,-6.25L27.042,75.969c-3.625,-3.646 -5.708,-8.677 -5.708,-13.802c0,-10.75 8.75,-19.5 19.5,-19.5h3.798C49.049,55.056 60.779,64 74.667,64H224c13.888,0 25.617,-8.944 30.035,-21.333h3.798c10.75,0 19.5,8.75 19.5,19.5c0,5.125 -2.083,10.156 -5.708,13.792L204.5,143.083c-3.979,3.969 -9.479,6.25 -15.083,6.25H109.25zM437.333,490.667H288c-5.875,0 -10.667,-4.781 -10.667,-10.667c0,-5.885 4.792,-10.667 10.667,-10.667h149.333c5.875,0 10.667,4.781 10.667,10.667C448,485.885 443.208,490.667 437.333,490.667zM471.167,469.333h-3.798C462.951,456.944 451.221,448 437.333,448H288c-13.888,0 -25.617,8.944 -30.035,21.333h-3.798c-10.75,0 -19.5,-8.75 -19.5,-19.5c0,-5.125 2.083,-10.156 5.708,-13.792l67.125,-67.125c3.979,-3.969 9.479,-6.25 15.083,-6.25h80.167c5.604,0 11.104,2.281 15.083,6.25l67.125,67.115c3.625,3.646 5.708,8.677 5.708,13.802C490.667,460.583 481.917,469.333 471.167,469.333z" />
 </vector>
