+++ app/src/main/res/drawable/ic_treadmill.xml
@@ -1,12 +1,12 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="70dp"
-    android:height="70dp"
+    android:width="50dp"
+    android:height="50dp"
     android:viewportWidth="512"
     android:viewportHeight="512">
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M373.813,394.675l-234.667,-10.667c-5.51,-0.135 -10.875,4.281 -11.135,10.177c-0.271,5.875 4.292,10.865 10.177,11.135l234.667,10.667c0.156,0.01 0.323,0.01 0.49,0.01c5.667,0 10.385,-4.458 10.646,-10.188C384.26,399.936 379.698,394.946 373.813,394.675z"/>
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M469.625,362.665l-361.007,-18.5L143.156,206.3c2.292,-9.146 10.167,-15.625 19.583,-16.115l157.813,-8.302c11.656,-0.615 20.781,-10.229 20.781,-21.896c0,-11.635 -9.094,-21.25 -20.708,-21.896l-167.927,-9.323c-5.688,-0.323 -11.115,-2.969 -14.865,-7.26L72.031,46.311c-3.521,-4.021 -9.469,-4.823 -13.948,-1.854l-32,21.333c-2.438,1.625 -4.104,4.188 -4.594,7.073c-0.5,2.885 0.219,5.854 1.979,8.198l53.646,71.531c4.323,5.771 5.448,13.26 3.01,20.031c0,0 -74.71,198.261 -76.829,203.907C1.177,382.177 0,388.272 0,394.665c0,25.754 18.353,47.298 42.667,52.254v11.746c0,5.896 4.771,10.667 10.667,10.667h64c5.896,0 10.667,-4.771 10.667,-10.667v-10.667h298.667v10.667c0,5.896 4.771,10.667 10.667,10.667h32c5.896,0 10.667,-4.771 10.667,-10.667v-12.181c18.352,-4.77 32,-21.326 32,-41.152C512,381.905 493.021,362.821 469.625,362.665zM53.333,426.665c-17.646,0 -32,-14.354 -32,-32c0,-17.646 14.354,-32 31.458,-32.01l0.891,0.046c17.482,0.193 31.651,14.439 31.651,31.965C85.333,412.311 70.979,426.665 53.333,426.665zM53.333,341.332c-4.076,0 -8.005,0.555 -11.822,1.424l58.686,-162.904c4.875,-13.563 2.625,-28.531 -6.021,-40.052L47.302,77.29l14.833,-9.896l59.635,68.156c7.51,8.594 18.354,13.875 29.74,14.51l167.927,10.51l-157.823,8.302c-18.833,1 -34.563,13.948 -39.146,32.24L86.91,343.052L53.333,341.332zM432.6,426.665H95.728c6.768,-8.941 10.939,-19.948 10.939,-32c0,-10.992 -3.352,-21.215 -9.074,-29.714l335.993,17.23c-4.34,6.678 -6.919,14.607 -6.919,23.15C426.667,413.138 428.926,420.364 432.6,426.665zM469.333,426.665c-11.76,0 -21.333,-9.573 -21.333,-21.333c0,-11.667 9.439,-21.13 21.07,-21.28c11.503,0.725 21.586,9.432 21.596,21.28C490.667,417.092 481.094,426.665 469.333,426.665z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M373.813,394.675l-234.667,-10.667c-5.51,-0.135 -10.875,4.281 -11.135,10.177c-0.271,5.875 4.292,10.865 10.177,11.135l234.667,10.667c0.156,0.01 0.323,0.01 0.49,0.01c5.667,0 10.385,-4.458 10.646,-10.188C384.26,399.936 379.698,394.946 373.813,394.675z" />
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M469.625,362.665l-361.007,-18.5L143.156,206.3c2.292,-9.146 10.167,-15.625 19.583,-16.115l157.813,-8.302c11.656,-0.615 20.781,-10.229 20.781,-21.896c0,-11.635 -9.094,-21.25 -20.708,-21.896l-167.927,-9.323c-5.688,-0.323 -11.115,-2.969 -14.865,-7.26L72.031,46.311c-3.521,-4.021 -9.469,-4.823 -13.948,-1.854l-32,21.333c-2.438,1.625 -4.104,4.188 -4.594,7.073c-0.5,2.885 0.219,5.854 1.979,8.198l53.646,71.531c4.323,5.771 5.448,13.26 3.01,20.031c0,0 -74.71,198.261 -76.829,203.907C1.177,382.177 0,388.272 0,394.665c0,25.754 18.353,47.298 42.667,52.254v11.746c0,5.896 4.771,10.667 10.667,10.667h64c5.896,0 10.667,-4.771 10.667,-10.667v-10.667h298.667v10.667c0,5.896 4.771,10.667 10.667,10.667h32c5.896,0 10.667,-4.771 10.667,-10.667v-12.181c18.352,-4.77 32,-21.326 32,-41.152C512,381.905 493.021,362.821 469.625,362.665zM53.333,426.665c-17.646,0 -32,-14.354 -32,-32c0,-17.646 14.354,-32 31.458,-32.01l0.891,0.046c17.482,0.193 31.651,14.439 31.651,31.965C85.333,412.311 70.979,426.665 53.333,426.665zM53.333,341.332c-4.076,0 -8.005,0.555 -11.822,1.424l58.686,-162.904c4.875,-13.563 2.625,-28.531 -6.021,-40.052L47.302,77.29l14.833,-9.896l59.635,68.156c7.51,8.594 18.354,13.875 29.74,14.51l167.927,10.51l-157.823,8.302c-18.833,1 -34.563,13.948 -39.146,32.24L86.91,343.052L53.333,341.332zM432.6,426.665H95.728c6.768,-8.941 10.939,-19.948 10.939,-32c0,-10.992 -3.352,-21.215 -9.074,-29.714l335.993,17.23c-4.34,6.678 -6.919,14.607 -6.919,23.15C426.667,413.138 428.926,420.364 432.6,426.665zM469.333,426.665c-11.76,0 -21.333,-9.573 -21.333,-21.333c0,-11.667 9.439,-21.13 21.07,-21.28c11.503,0.725 21.586,9.432 21.596,21.28C490.667,417.092 481.094,426.665 469.333,426.665z" />
 </vector>
