+++ app/src/main/res/drawable/ic_workout.xml
@@ -1,12 +1,12 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="70dp"
-    android:height="70dp"
+    android:width="50dp"
+    android:height="50dp"
     android:viewportWidth="512"
     android:viewportHeight="512">
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M444.875,348.875c4.167,-4.167 4.167,-10.917 0,-15.083l-21.333,-21.333c-4.167,-4.167 -10.917,-4.167 -15.083,0l-5.842,5.842c-11.816,-8.348 -25.284,-14.488 -39.949,-17.478V288V96c0,-2.833 -1.125,-5.542 -3.125,-7.542L274.208,3.125c-2,-2 -4.708,-3.125 -7.542,-3.125h-160C83.146,0 64,19.135 64,42.667v320c0,23.531 19.146,42.667 42.667,42.667h128c0,58.813 47.854,106.667 106.667,106.667C400.146,512 448,464.146 448,405.333c0,-17.255 -4.508,-33.331 -11.815,-47.768L444.875,348.875zM277.333,36.417l48.917,48.917h-27.583c-11.771,0 -21.333,-9.573 -21.333,-21.333V36.417zM236.823,384H106.667c-11.771,0 -21.333,-9.573 -21.333,-21.333v-320c0,-11.76 9.563,-21.333 21.333,-21.333H256V64c0,23.531 19.146,42.667 42.667,42.667h42.667v170.667h-10.667c-5.896,0 -10.667,4.771 -10.667,10.667v12.823c-14.665,2.99 -28.133,9.13 -39.949,17.478l-5.842,-5.842c-4.167,-4.167 -10.917,-4.167 -15.083,0l-21.333,21.333c-4.167,4.167 -4.167,10.917 0,15.083l8.69,8.69C242.268,365.891 238.738,374.605 236.823,384zM341.333,490.667c-47.063,0 -85.333,-38.281 -85.333,-85.333C256,358.281 294.271,320 341.333,320c47.063,0 85.333,38.281 85.333,85.333C426.667,452.385 388.396,490.667 341.333,490.667z"/>
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M354.083,341.542c-5.521,-1.188 -11.375,2.573 -12.542,8.365l-7.098,35.488C326.092,388.29 320,396.013 320,405.333c0,11.76 9.563,21.333 21.333,21.333c11.771,0 21.333,-9.573 21.333,-21.333c0,-6.335 -2.9,-11.872 -7.299,-15.784l7.091,-35.456C363.625,348.313 359.875,342.698 354.083,341.542z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M444.875,348.875c4.167,-4.167 4.167,-10.917 0,-15.083l-21.333,-21.333c-4.167,-4.167 -10.917,-4.167 -15.083,0l-5.842,5.842c-11.816,-8.348 -25.284,-14.488 -39.949,-17.478V288V96c0,-2.833 -1.125,-5.542 -3.125,-7.542L274.208,3.125c-2,-2 -4.708,-3.125 -7.542,-3.125h-160C83.146,0 64,19.135 64,42.667v320c0,23.531 19.146,42.667 42.667,42.667h128c0,58.813 47.854,106.667 106.667,106.667C400.146,512 448,464.146 448,405.333c0,-17.255 -4.508,-33.331 -11.815,-47.768L444.875,348.875zM277.333,36.417l48.917,48.917h-27.583c-11.771,0 -21.333,-9.573 -21.333,-21.333V36.417zM236.823,384H106.667c-11.771,0 -21.333,-9.573 -21.333,-21.333v-320c0,-11.76 9.563,-21.333 21.333,-21.333H256V64c0,23.531 19.146,42.667 42.667,42.667h42.667v170.667h-10.667c-5.896,0 -10.667,4.771 -10.667,10.667v12.823c-14.665,2.99 -28.133,9.13 -39.949,17.478l-5.842,-5.842c-4.167,-4.167 -10.917,-4.167 -15.083,0l-21.333,21.333c-4.167,4.167 -4.167,10.917 0,15.083l8.69,8.69C242.268,365.891 238.738,374.605 236.823,384zM341.333,490.667c-47.063,0 -85.333,-38.281 -85.333,-85.333C256,358.281 294.271,320 341.333,320c47.063,0 85.333,38.281 85.333,85.333C426.667,452.385 388.396,490.667 341.333,490.667z" />
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M354.083,341.542c-5.521,-1.188 -11.375,2.573 -12.542,8.365l-7.098,35.488C326.092,388.29 320,396.013 320,405.333c0,11.76 9.563,21.333 21.333,21.333c11.771,0 21.333,-9.573 21.333,-21.333c0,-6.335 -2.9,-11.872 -7.299,-15.784l7.091,-35.456C363.625,348.313 359.875,342.698 354.083,341.542z" />
 </vector>
