+++ app/src/main/res/drawable/ic_workout_3.xml
@@ -1,9 +1,9 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="70dp"
-    android:height="70dp"
+    android:width="50dp"
+    android:height="50dp"
     android:viewportWidth="469.333"
     android:viewportHeight="469.333">
-  <path
-      android:fillColor="#FF000000"
-      android:pathData="M437.333,85.333c-17.646,0 -32,14.354 -32,32V128H320V32c0,-17.646 -14.354,-32 -32,-32c-17.646,0 -32,14.354 -32,32v10.667H64V32C64,14.354 49.646,0 32,0C14.354,0 0,14.354 0,32v426.667c0,5.896 4.771,10.667 10.667,10.667h42.667c5.896,0 10.667,-4.771 10.667,-10.667v-352h42.667v65.965c-12.389,4.418 -21.333,16.147 -21.333,30.035c0,17.646 14.354,32 32,32c17.646,0 32,-14.354 32,-32c0,-13.888 -8.944,-25.617 -21.333,-30.035v-65.965h64v65.965c-12.389,4.418 -21.333,16.147 -21.333,30.035c0,17.646 14.354,32 32,32s32,-14.354 32,-32c0,-13.888 -8.944,-25.617 -21.333,-30.035v-65.965H256v352c0,5.896 4.771,10.667 10.667,10.667h42.667c5.896,0 10.667,-4.771 10.667,-10.667V192h85.333v266.667c0,5.896 4.771,10.667 10.667,10.667h42.667c5.896,0 10.667,-4.771 10.667,-10.667V117.333C469.333,99.687 454.979,85.333 437.333,85.333zM53.333,85.333c-5.896,0 -10.667,4.771 -10.667,10.667v352H21.333V32c0,-5.885 4.781,-10.667 10.667,-10.667c5.885,0 10.667,4.781 10.667,10.667v21.333C42.667,59.229 47.437,64 53.333,64h53.333v21.333H53.333zM117.333,213.333c-5.885,0 -10.667,-4.781 -10.667,-10.667S111.448,192 117.333,192c5.885,0 10.667,4.781 10.667,10.667S123.219,213.333 117.333,213.333zM128,85.333V64h64v21.333H128zM202.667,213.333c-5.885,0 -10.667,-4.781 -10.667,-10.667S196.781,192 202.667,192s10.667,4.781 10.667,10.667S208.552,213.333 202.667,213.333zM448,448h-21.333V181.333c0,-5.896 -4.771,-10.667 -10.667,-10.667H309.333c-5.896,0 -10.667,4.771 -10.667,10.667V448h-21.333V96c0,-5.896 -4.771,-10.667 -10.667,-10.667h-53.333V64h53.333c5.896,0 10.667,-4.771 10.667,-10.667V32c0,-5.885 4.781,-10.667 10.667,-10.667c5.885,0 10.667,4.781 10.667,10.667v106.667c0,5.896 4.771,10.667 10.667,10.667H416c5.896,0 10.667,-4.771 10.667,-10.667v-21.333c0,-5.885 4.781,-10.667 10.667,-10.667S448,111.448 448,117.333V448z"/>
+    <path
+        android:fillColor="#FF000000"
+        android:pathData="M437.333,85.333c-17.646,0 -32,14.354 -32,32V128H320V32c0,-17.646 -14.354,-32 -32,-32c-17.646,0 -32,14.354 -32,32v10.667H64V32C64,14.354 49.646,0 32,0C14.354,0 0,14.354 0,32v426.667c0,5.896 4.771,10.667 10.667,10.667h42.667c5.896,0 10.667,-4.771 10.667,-10.667v-352h42.667v65.965c-12.389,4.418 -21.333,16.147 -21.333,30.035c0,17.646 14.354,32 32,32c17.646,0 32,-14.354 32,-32c0,-13.888 -8.944,-25.617 -21.333,-30.035v-65.965h64v65.965c-12.389,4.418 -21.333,16.147 -21.333,30.035c0,17.646 14.354,32 32,32s32,-14.354 32,-32c0,-13.888 -8.944,-25.617 -21.333,-30.035v-65.965H256v352c0,5.896 4.771,10.667 10.667,10.667h42.667c5.896,0 10.667,-4.771 10.667,-10.667V192h85.333v266.667c0,5.896 4.771,10.667 10.667,10.667h42.667c5.896,0 10.667,-4.771 10.667,-10.667V117.333C469.333,99.687 454.979,85.333 437.333,85.333zM53.333,85.333c-5.896,0 -10.667,4.771 -10.667,10.667v352H21.333V32c0,-5.885 4.781,-10.667 10.667,-10.667c5.885,0 10.667,4.781 10.667,10.667v21.333C42.667,59.229 47.437,64 53.333,64h53.333v21.333H53.333zM117.333,213.333c-5.885,0 -10.667,-4.781 -10.667,-10.667S111.448,192 117.333,192c5.885,0 10.667,4.781 10.667,10.667S123.219,213.333 117.333,213.333zM128,85.333V64h64v21.333H128zM202.667,213.333c-5.885,0 -10.667,-4.781 -10.667,-10.667S196.781,192 202.667,192s10.667,4.781 10.667,10.667S208.552,213.333 202.667,213.333zM448,448h-21.333V181.333c0,-5.896 -4.771,-10.667 -10.667,-10.667H309.333c-5.896,0 -10.667,4.771 -10.667,10.667V448h-21.333V96c0,-5.896 -4.771,-10.667 -10.667,-10.667h-53.333V64h53.333c5.896,0 10.667,-4.771 10.667,-10.667V32c0,-5.885 4.781,-10.667 10.667,-10.667c5.885,0 10.667,4.781 10.667,10.667v106.667c0,5.896 4.771,10.667 10.667,10.667H416c5.896,0 10.667,-4.771 10.667,-10.667v-21.333c0,-5.885 4.781,-10.667 10.667,-10.667S448,111.448 448,117.333V448z" />
 </vector>
